organized directory and fixed according to changes
