+++ app/src/main/java/com/example/smartalarm/AddEventActivity.java
@@ -10,6 +10,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.DialogFragment;
 import com.example.smartalarm.event.DelayedEvent;
 import com.example.smartalarm.event.Event;
+import com.example.smartalarm.fragment.DatePickerFragment;
+import com.example.smartalarm.fragment.TimePickerFragment;
 
 import java.io.Serializable;
 import java.util.Calendar;
