+++ app/src/main/java/com/example/smartalarm/ViewEventActivity.java
@@ -9,6 +9,8 @@ import android.widget.*;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.DialogFragment;
 import com.example.smartalarm.event.Event;
+import com.example.smartalarm.fragment.DatePickerFragment;
+import com.example.smartalarm.fragment.TimePickerFragment;
 
 import java.util.Calendar;
