+++ app/src/main/java/com/example/smartalarm/adapter/ConcreteEventListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.smartalarm;
+package com.example.smartalarm.adapter;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -8,6 +8,9 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.smartalarm.R;
+import com.example.smartalarm.event.Event;
 
 import java.util.List;
 
@@ -34,7 +37,9 @@ public class ConcreteEventListAdapter extends
             int itemPosition = getLayoutPosition();
 
             Event event = mEvents.get(itemPosition);
-            event.activated = !event.activated;
+            if (event.getCurrentState() == "activated") { event.deactivateEvent(); }
+            if (event.getCurrentState() == "deactivated") { event.activateEvent(); }
+
             mAdapter.notifyDataSetChanged();
             // TODO: Set On of Off
         }
@@ -58,7 +63,7 @@ public class ConcreteEventListAdapter extends
     @Override
     public void onBindViewHolder(@NonNull ConcreteEventViewHolder holder, int position) {
         // Retrieve the data for that position.
-        String mCurrent = mEvents.get(position).name;
+        String mCurrent = mEvents.get(position).getName();
         // Add the data to the view holder.
         holder.mItemView.setText(mCurrent);
     }
