+++ app/src/main/java/com/example/smartalarm/event/DelayedEvent.java
@@ -38,6 +38,9 @@ public class DelayedEvent extends Event implements PropertyChangeListener {
    }
 
    @Override
+   public String getCurrentState(){ return currentState; };
+
+   @Override
    public void setDelay(Calendar c) {
       delay = c;
    }
